properly fix gradients offsets for scrollpanes

+++ yeehaw/src/organizer.rs
@@ -347,21 +347,13 @@ impl ElementOrganizer {
                         // currently using rayon for parallelization
                         if *details.overflow.borrow() {
                             dcps.par_iter_mut().for_each(|dcp| {
-                                dcp.set_draw_size_offset_colors(
-                                    child_s,
-                                    start_x as u16,
-                                    start_y as u16,
-                                );
+                                dcp.set_draw_size_offset_colors(child_s, start_x, start_y);
                                 dcp.x += start_x as u16;
                                 dcp.y += start_y as u16;
                             });
                         } else {
                             dcps.par_iter_mut().for_each(|dcp| {
-                                dcp.set_draw_size_offset_colors(
-                                    child_s,
-                                    start_x as u16,
-                                    start_y as u16,
-                                );
+                                dcp.set_draw_size_offset_colors(child_s, start_x, start_y);
                                 if dcp.x >= child_s.width || dcp.y >= child_s.height {
                                     // it'd be better to delete, but we can't delete from a parallel iterator
                                     // also using a filter here its slower that this
